+++ NEWS.md
@@ -1,4 +1,15 @@
 
+'replyr' 0.3.01 2017-05-13
+
+ * much better tracking of temp-handle names.
+ * more tests/checks.
+ 
+'replyr' 0.2.6 2017-05-08
+
+ * use dplyr::ungroup a few places.
+ * add replyr::expandColumn.
+ * deal with logical and list columns in summary.
+ 
 'replyr' 0.2.5 2017-04-14
 
  * Remove all use of dplyr underbar/underscore forms to future proof as these are already deprecated in dev ( https://github.com/hadley/dplyr/blob/master/R/manip.r 33d0b1ecaa22644e0d78345b13d5f8aca5d949b ).